requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
matplotlib>=3.2.2
Pillow>=8.0.0
PyYAML>=5.3.1
tqdm>=4.41.0
streamlit>=1.11.0
